feat: implement hosted login flows with embedded login, social login, and error handling
