+++ Sources/FronteggSwift/FronteggAuth.swift
@@ -60,11 +60,12 @@ public class FronteggAuth: FronteggState {
     var loginCompletion: CompletionHandler? = nil
     // internal for extension access (Connectivity, Testing, SessionRestore)
     var networkMonitoringToken: NetworkStatusMonitor.OnChangeToken?
-    private var pendingOAuthErrorContext: FronteggOAuthErrorContext?
-    private var pendingOAuthErrorPresentationWorkItem: DispatchWorkItem?
-    private var pendingEmbeddedOAuthErrorFallbackWorkItem: DispatchWorkItem?
-    private let oauthErrorPresentationDelay: TimeInterval = 0.35
-    private let embeddedOAuthErrorRecoveryFallbackDelay: TimeInterval = 1.25
+    // internal for extension access (FronteggAuth+OAuthErrors.swift)
+    var pendingOAuthErrorContext: FronteggOAuthErrorContext?
+    var pendingOAuthErrorPresentationWorkItem: DispatchWorkItem?
+    var pendingEmbeddedOAuthErrorFallbackWorkItem: DispatchWorkItem?
+    let oauthErrorPresentationDelay: TimeInterval = 0.35
+    let embeddedOAuthErrorRecoveryFallbackDelay: TimeInterval = 1.25
     // internal for extension access (SessionRestore, Testing, Refresh, HostedFlows)
     var isInitializingWithTokens: Bool = false
     @MainActor var isLoginInProgress: Bool = false
@@ -646,7 +647,7 @@ public class FronteggAuth: FronteggState {
     }
 
 
-    private func resolveBestEffortUser(accessToken: String, includeInMemory: Bool = true) -> User? {
+    func resolveBestEffortUser(accessToken: String, includeInMemory: Bool = true) -> User? {
         if includeInMemory, let currentUser = self.user {
             return currentUser
         }
@@ -671,7 +672,7 @@ public class FronteggAuth: FronteggState {
         return nil
     }
 
-    private func logPreservedSessionAfterAuthoritativeUserLoadFailure(_ error: Error, stage: String) {
+    func logPreservedSessionAfterAuthoritativeUserLoadFailure(_ error: Error, stage: String) {
         let reflectedError = String(reflecting: error)
         let connectivityFailure = isConnectivityError(error)
         if connectivityFailure {
@@ -705,7 +706,7 @@ public class FronteggAuth: FronteggState {
         }
     }
 
-    private func setCredentialsInternal(accessToken: String, refreshToken: String, user: User? = nil, hydrationMode: CredentialHydrationMode) async {
+    func setCredentialsInternal(accessToken: String, refreshToken: String, user: User? = nil, hydrationMode: CredentialHydrationMode) async {
         self.logger.info("Setting credentials (refresh token length: \(refreshToken.count), hydrationMode: \(hydrationModeDescription(hydrationMode)))")
 
         var accessToken = accessToken
@@ -1128,7 +1129,7 @@ public class FronteggAuth: FronteggState {
     /// Checks network path availability using NWPathMonitor with a timeout.
     /// Path checks are advisory only in restricted-network environments (e.g., aircraft with whitelisted domains).
     /// Actual HTTP request failures remain the authoritative source of truth.
-    private func checkNetworkPath(timeout: UInt64 = 500_000_000) async -> Bool {
+    func checkNetworkPath(timeout: UInt64 = 500_000_000) async -> Bool {
 #if DEBUG
         if let override = Self.testNetworkPathAvailabilityOverride {
             return override
@@ -1170,7 +1171,7 @@ public class FronteggAuth: FronteggState {
 
     /// Resolves stored session artifacts (tokens, offline user, tenant) using consistent precedence:
     /// lastActiveTenantId → user.activeTenant → offlineUser.activeTenant → legacy global tokens.
-    private func resolveStoredSessionArtifacts(enableSessionPerTenant: Bool) -> StoredSessionArtifacts {
+    func resolveStoredSessionArtifacts(enableSessionPerTenant: Bool) -> StoredSessionArtifacts {
         var refreshToken: String? = nil
         var accessToken: String? = nil
         var tenantId: String? = nil
@@ -1211,7 +1212,7 @@ public class FronteggAuth: FronteggState {
 
     /// Shared state update for connectivity loss on manual refresh paths.
     /// Does NOT enqueue retries — scheduled refresh paths handle their own retry/backoff logic.
-    private func applyConnectivityLossState(enableOfflineMode: Bool) async {
+    func applyConnectivityLossState(enableOfflineMode: Bool) async {
         let hasTokens = (self.refreshToken != nil || self.accessToken != nil)
         guard hasTokens else { return }
 
@@ -1240,7 +1241,7 @@ public class FronteggAuth: FronteggState {
 
     /// Manual refresh failures should preserve the cached session and rely on reconnect monitoring
     /// instead of spinning scheduled refresh retries against a disconnected or blocked network.
-    private func startOfflineMonitoringAfterManualConnectivityFailure(enableOfflineMode: Bool) {
+    func startOfflineMonitoringAfterManualConnectivityFailure(enableOfflineMode: Bool) {
         guard enableOfflineMode else { return }
 
         let hasTokens = (self.refreshToken != nil || self.accessToken != nil)
@@ -1333,7 +1334,7 @@ public class FronteggAuth: FronteggState {
     
     /// Starts network monitoring so that `reconnectedToInternet()` fires when connectivity returns.
     /// Safe to call multiple times — stops existing monitoring first to avoid duplicates.
-    private func ensureOfflineMonitoringActive(intervalOverride: TimeInterval? = nil, emitInitialState: Bool = true) {
+    func ensureOfflineMonitoringActive(intervalOverride: TimeInterval? = nil, emitInitialState: Bool = true) {
         let config = try? PlistHelper.fronteggConfig()
         let monitoringInterval = intervalOverride ?? config?.networkMonitoringInterval ?? 10
 
@@ -1748,2057 +1749,12 @@ public class FronteggAuth: FronteggState {
     }
     
     
-    func handleHostedLoginCallback(_ code: String, _ codeVerifier: String, _ completion: @escaping FronteggAuth.CompletionHandler) {
-        handleHostedLoginCallback(
-            code,
-            codeVerifier,
-            oauthState: nil,
-            redirectUri: nil,
-            flow: .login,
-            completePendingFlowOnSuccess: false,
-            completion: completion
-        )
-    }
+    // MARK: OAuth Callbacks — see FronteggAuth+OAuthCallbacks.swift
+
     
-    func handleHostedLoginCallback(_ code: String, _ codeVerifier: String?, _ completion: @escaping FronteggAuth.CompletionHandler) {
-        handleHostedLoginCallback(
-            code,
-            codeVerifier,
-            oauthState: nil,
-            redirectUri: nil,
-            flow: .login,
-            completePendingFlowOnSuccess: false,
-            completion: completion
-        )
-    }
-    
-    func handleHostedLoginCallback(_ code: String, _ codeVerifier: String?, redirectUri: String?, _ completion: @escaping FronteggAuth.CompletionHandler) {
-        handleHostedLoginCallback(
-            code,
-            codeVerifier,
-            oauthState: nil,
-            redirectUri: redirectUri,
-            flow: .login,
-            completePendingFlowOnSuccess: false,
-            completion: completion
-        )
-    }
-
-    private func normalizedOAuthMessageComponent(_ value: String?) -> String? {
-        guard let value else {
-            return nil
-        }
-
-        let normalizedValue = value
-            .replacingOccurrences(of: "+", with: " ")
-            .trimmingCharacters(in: .whitespacesAndNewlines)
-
-        guard !normalizedValue.isEmpty else {
-            return nil
-        }
-
-        return normalizedValue
-    }
-
-    private func oauthDisplayMessage(
-        errorCode: String?,
-        errorDescription: String?,
-        fallbackMessage: String? = nil
-    ) -> String {
-        oauthDisplayMessage(
-            normalizedErrorCode: normalizedOAuthMessageComponent(errorCode),
-            normalizedErrorDescription: normalizedOAuthMessageComponent(errorDescription),
-            normalizedFallbackMessage: normalizedOAuthMessageComponent(fallbackMessage)
-        )
-    }
-
-    private func oauthDisplayMessage(
-        normalizedErrorCode: String?,
-        normalizedErrorDescription: String?,
-        normalizedFallbackMessage: String? = nil
-    ) -> String {
-        if let normalizedErrorCode, let normalizedErrorDescription {
-            return "\(normalizedErrorCode): \(normalizedErrorDescription)"
-        }
-
-        if let normalizedErrorDescription {
-            return normalizedErrorDescription
-        }
-
-        if let normalizedErrorCode {
-            return normalizedErrorCode
-        }
-
-        if let normalizedFallbackMessage {
-            return normalizedFallbackMessage
-        }
-
-        return FronteggError.authError(.unknown).localizedDescription
-    }
-
-    func oauthFailureDetails(
-        errorCode: String?,
-        errorDescription: String?,
-        fallbackError: FronteggError? = nil
-    ) -> OAuthFailureDetails {
-        oauthFailureDetails(
-            normalizedErrorCode: normalizedOAuthMessageComponent(errorCode),
-            normalizedErrorDescription: normalizedOAuthMessageComponent(errorDescription),
-            fallbackError: fallbackError
-        )
-    }
-
-    private func oauthFailureDetails(
-        normalizedErrorCode: String?,
-        normalizedErrorDescription: String?,
-        fallbackError: FronteggError? = nil
-    ) -> OAuthFailureDetails {
-        let message = oauthDisplayMessage(
-            normalizedErrorCode: normalizedErrorCode,
-            normalizedErrorDescription: normalizedErrorDescription,
-            normalizedFallbackMessage: normalizedOAuthMessageComponent(
-                fallbackError?.localizedDescription
-            )
-        )
-
-        if normalizedErrorCode != nil || normalizedErrorDescription != nil {
-            return OAuthFailureDetails(
-                error: FronteggError.authError(.oauthError(message)),
-                errorCode: normalizedErrorCode,
-                errorDescription: normalizedErrorDescription
-            )
-        }
-
-        return OAuthFailureDetails(
-            error: fallbackError ?? FronteggError.authError(.oauthError(message)),
-            errorCode: nil,
-            errorDescription: nil
-        )
-    }
-
-    func oauthFailureDetails(from queryItems: [String: String]) -> OAuthFailureDetails? {
-        let errorCode = normalizedOAuthMessageComponent(queryItems["error"])
-        let errorDescription = normalizedOAuthMessageComponent(queryItems["error_description"])
-
-        guard errorCode != nil || errorDescription != nil else {
-            return nil
-        }
-
-        return oauthFailureDetails(
-            normalizedErrorCode: errorCode,
-            normalizedErrorDescription: errorDescription
-        )
-    }
-
-    private func shouldSuppressOAuthErrorPresentation(for error: FronteggError) -> Bool {
-        switch error {
-        case .authError(let authError):
-            switch authError {
-            case .operationCanceled:
-                return true
-            case .other(let underlyingError):
-                return Self.isUserCancelledOAuthFlow(underlyingError)
-            default:
-                return false
-            }
-        default:
-            return false
-        }
-    }
-
-    internal func reportOAuthFailure(
-        error: FronteggError,
-        flow: FronteggOAuthFlow,
-        errorCode: String? = nil,
-        errorDescription: String? = nil,
-        embeddedMode: Bool? = nil
-    ) {
-        guard !shouldSuppressOAuthErrorPresentation(for: error) else {
-            return
-        }
-
-        enqueueOAuthFailurePresentation(
-            error: error,
-            flow: flow,
-            normalizedErrorCode: normalizedOAuthMessageComponent(errorCode),
-            normalizedErrorDescription: normalizedOAuthMessageComponent(errorDescription),
-            normalizedFallbackMessage: normalizedOAuthMessageComponent(error.localizedDescription),
-            embeddedMode: embeddedMode
-        )
-    }
-
-    internal func reportOAuthFailure(
-        details: OAuthFailureDetails,
-        flow: FronteggOAuthFlow,
-        embeddedMode: Bool? = nil
-    ) {
-        guard !shouldSuppressOAuthErrorPresentation(for: details.error) else {
-            return
-        }
-
-        enqueueOAuthFailurePresentation(
-            error: details.error,
-            flow: flow,
-            normalizedErrorCode: details.errorCode,
-            normalizedErrorDescription: details.errorDescription,
-            normalizedFallbackMessage: normalizedOAuthMessageComponent(
-                details.error.localizedDescription
-            ),
-            embeddedMode: embeddedMode
-        )
-    }
-
-    private func enqueueOAuthFailurePresentation(
-        error: FronteggError,
-        flow: FronteggOAuthFlow,
-        normalizedErrorCode: String?,
-        normalizedErrorDescription: String?,
-        normalizedFallbackMessage: String?,
-        embeddedMode: Bool?
-    ) {
-        let context = FronteggOAuthErrorContext(
-            displayMessage: oauthDisplayMessage(
-                normalizedErrorCode: normalizedErrorCode,
-                normalizedErrorDescription: normalizedErrorDescription,
-                normalizedFallbackMessage: normalizedFallbackMessage
-            ),
-            errorCode: normalizedErrorCode,
-            errorDescription: normalizedErrorDescription,
-            error: error,
-            flow: flow,
-            embeddedMode: embeddedMode ?? self.embeddedMode
-        )
-
-        Task { @MainActor in
-            FronteggRuntime.testingLog(
-                "E2E queued OAuth error flow=\(context.flow) embedded=\(context.embeddedMode) code=\(context.errorCode ?? "nil") message=\(context.displayMessage)"
-            )
-            self.pendingOAuthErrorContext = context
-            let shouldDeferEmbeddedPresentation = context.embeddedMode && self.webview != nil
-            if shouldDeferEmbeddedPresentation {
-                self.pendingOAuthErrorPresentationWorkItem?.cancel()
-                self.pendingOAuthErrorPresentationWorkItem = nil
-                self.scheduleEmbeddedOAuthErrorFallbackIfNeeded(for: context)
-                return
-            }
-
-            self.pendingEmbeddedOAuthErrorFallbackWorkItem?.cancel()
-            self.pendingEmbeddedOAuthErrorFallbackWorkItem = nil
-            self.flushPendingOAuthErrorPresentationIfNeeded()
-        }
-    }
-
-    @MainActor
-    internal func flushPendingOAuthErrorPresentationIfNeeded(delayIfNeeded: Bool = false) {
-        let requiresForegroundWindow =
-            FronteggOAuthErrorRuntimeSettings.presentation == .toast
-
-        guard !requiresForegroundWindow || UIApplication.shared.applicationState != .background else {
-            return
-        }
-
-        guard let context = pendingOAuthErrorContext else {
-            pendingEmbeddedOAuthErrorFallbackWorkItem?.cancel()
-            pendingEmbeddedOAuthErrorFallbackWorkItem = nil
-            return
-        }
-
-        pendingEmbeddedOAuthErrorFallbackWorkItem?.cancel()
-        pendingEmbeddedOAuthErrorFallbackWorkItem = nil
-        pendingOAuthErrorContext = nil
-        FronteggRuntime.testingLog(
-            "E2E flushing OAuth error flow=\(context.flow) embedded=\(context.embeddedMode) delay=\(delayIfNeeded)"
-        )
-        scheduleOAuthErrorPresentation(
-            context,
-            delay: delayIfNeeded ? oauthErrorPresentationDelay : 0
-        )
-    }
-
-    @MainActor
-    private func scheduleOAuthErrorPresentation(
-        _ context: FronteggOAuthErrorContext,
-        delay: TimeInterval
-    ) {
-        pendingOAuthErrorPresentationWorkItem?.cancel()
-
-        let shouldPresentImmediately =
-            delay <= 0 && FronteggOAuthErrorRuntimeSettings.presentation == .delegate
-        if shouldPresentImmediately {
-            pendingOAuthErrorPresentationWorkItem = nil
-            presentOAuthError(context)
-            return
-        }
-
-        let workItem = DispatchWorkItem { [weak self] in
-            guard let self else { return }
-            self.pendingOAuthErrorPresentationWorkItem = nil
-            self.presentOAuthError(context)
-        }
-
-        pendingOAuthErrorPresentationWorkItem = workItem
-        if delay > 0 {
-            DispatchQueue.main.asyncAfter(deadline: .now() + delay, execute: workItem)
-        } else {
-            DispatchQueue.main.async(execute: workItem)
-        }
-    }
-
-    @MainActor
-    private func scheduleEmbeddedOAuthErrorFallbackIfNeeded(for context: FronteggOAuthErrorContext) {
-        pendingEmbeddedOAuthErrorFallbackWorkItem?.cancel()
-
-        let workItem = DispatchWorkItem { [weak self] in
-            guard let self else { return }
-            self.pendingEmbeddedOAuthErrorFallbackWorkItem = nil
-            guard let pendingContext = self.pendingOAuthErrorContext else {
-                return
-            }
-
-            let isSamePendingError =
-                pendingContext.displayMessage == context.displayMessage &&
-                pendingContext.errorCode == context.errorCode &&
-                pendingContext.errorDescription == context.errorDescription &&
-                pendingContext.flow == context.flow &&
-                pendingContext.embeddedMode == context.embeddedMode
-
-            guard isSamePendingError else {
-                return
-            }
-
-            self.logger.warning("Embedded OAuth error recovery did not settle in time. Presenting pending OAuth error and clearing the loader.")
-            FronteggRuntime.testingLog(
-                "E2E embedded OAuth error fallback firing flow=\(context.flow) code=\(context.errorCode ?? "nil")"
-            )
-            self.setWebLoading(false)
-            self.flushPendingOAuthErrorPresentationIfNeeded()
-        }
-
-        pendingEmbeddedOAuthErrorFallbackWorkItem = workItem
-        DispatchQueue.main.asyncAfter(
-            deadline: .now() + embeddedOAuthErrorRecoveryFallbackDelay,
-            execute: workItem
-        )
-    }
-
-    @MainActor
-    private func presentOAuthError(_ context: FronteggOAuthErrorContext) {
-        FronteggRuntime.testingLog(
-            "E2E presenting OAuth error flow=\(context.flow) embedded=\(context.embeddedMode) message=\(context.displayMessage)"
-        )
-        switch FronteggOAuthErrorRuntimeSettings.presentation {
-        case .toast:
-            let window = self.resolveOAuthErrorPresentationWindow()
-            if window == nil {
-                self.logger.warning("OAuth failure toast could not find a presentation window")
-            }
-            FronteggOAuthToastPresenter.shared.show(message: context.displayMessage, in: window)
-        case .delegate:
-            FronteggOAuthErrorRuntimeSettings.delegateBox.value?.fronteggSDK(didReceiveOAuthError: context)
-        }
-    }
-
-    @MainActor
-    private func resolveOAuthErrorPresentationWindow() -> UIWindow? {
-        if let window = self.webview?.window {
-            return window
-        }
-
-        if let window = VCHolder.shared.vc?.presentedViewController?.view.window {
-            return window
-        }
-
-        if let window = VCHolder.shared.vc?.view.window {
-            return window
-        }
-
-        if let window = self.getRootVC(true)?.view.window {
-            return window
-        }
-
-        if let window = self.getRootVC()?.view.window {
-            return window
-        }
-
-        return UIWindow.fronteggPresentationCandidate
-    }
-
-    private func completePendingOAuthFlowIfNeeded(
-        oauthState: String?,
-        codeVerifier: String?,
-        matchedPendingOAuthState: Bool,
-        completePendingFlowOnSuccess: Bool
-    ) {
-        guard completePendingFlowOnSuccess else { return }
-        if matchedPendingOAuthState, let oauthState, !oauthState.isEmpty {
-            CredentialManager.clearPendingOAuth(state: oauthState)
-            CredentialManager.clearCodeVerifierIfMatching(codeVerifier)
-            return
-        }
-        CredentialManager.completePendingOAuthFlow(using: codeVerifier)
-    }
-
-    private func clearMatchedPendingOAuthFlowIfNeeded(
-        oauthState: String?,
-        codeVerifier: String?,
-        matchedPendingOAuthState: Bool,
-        managePendingOAuthFlow: Bool
-    ) {
-        guard managePendingOAuthFlow,
-              matchedPendingOAuthState,
-              let oauthState,
-              !oauthState.isEmpty else {
-            return
-        }
-
-        CredentialManager.clearPendingOAuth(state: oauthState)
-        CredentialManager.clearCodeVerifierIfMatching(codeVerifier)
-    }
-
-    private func oauthCodeVerifierError(
-        for oauthState: String?,
-        resolution: CredentialManager.CodeVerifierResolution
-    ) -> FronteggError.Authentication {
-        if oauthState != nil && resolution.hasPendingOAuthStates {
-            return .invalidOAuthState
-        }
-        return .codeVerifierNotFound
-    }
-
-    func handleHostedLoginCallback(
-        _ code: String,
-        _ codeVerifier: String?,
-        oauthState: String?,
-        redirectUri: String?,
-        flow: FronteggOAuthFlow = .login,
-        completePendingFlowOnSuccess: Bool,
-        matchedPendingOAuthState: Bool = false,
-        completion: @escaping FronteggAuth.CompletionHandler
-    ) {
-
-        Task {
-            // Use provided redirectUri or generate default one.
-            // For magic link flow, use the redirectUri from the callback URL.
-            let redirectUri = redirectUri ?? generateRedirectUri()
-
-            await MainActor.run {
-                self.setIsLoading(true)
-                self.isLoginInProgress = true
-            }
-
-            self.logger.info("Handling hosted login callback (redirectUri: \(redirectUri), hasCodeVerifier: \(codeVerifier != nil))")
-            logger.info("Exchanging hosted login callback token")
-            let (responseData, error) = await api.exchangeToken(
-                code: code,
-                redirectUrl: redirectUri,
-                codeVerifier: codeVerifier
-            )
-            
-            if let error {
-                logger.error("Hosted login token exchange failed: \(error.localizedDescription)")
-                
-                SentryHelper.logError(error, context: [
-                    "token_exchange": [
-                        "redirectUri": redirectUri,
-                        "codeLength": code.count,
-                        "codeVerifierUsed": codeVerifier != nil,
-                        "method": "handleHostedLoginCallback"
-                    ],
-                    "error": [
-                        "type": "token_exchange_failed",
-                        "stage": "api_exchange"
-                    ]
-                ])
-
-                self.clearMatchedPendingOAuthFlowIfNeeded(
-                    oauthState: oauthState,
-                    codeVerifier: codeVerifier,
-                    matchedPendingOAuthState: matchedPendingOAuthState,
-                    managePendingOAuthFlow: completePendingFlowOnSuccess
-                )
-                self.reportOAuthFailure(error: error, flow: flow)
-                await MainActor.run {
-                    self.isLoginInProgress = false
-                    completion(.failure(error))
-                    self.setIsLoading(false)
-                    self.setWebLoading(false)
-                }
-                return
-            }
-            
-            guard let data = responseData else {
-                logger.error("Hosted login token exchange returned nil response data")
-                let authError = FronteggError.authError(.failedToAuthenticate)
-                self.clearMatchedPendingOAuthFlowIfNeeded(
-                    oauthState: oauthState,
-                    codeVerifier: codeVerifier,
-                    matchedPendingOAuthState: matchedPendingOAuthState,
-                    managePendingOAuthFlow: completePendingFlowOnSuccess
-                )
-                self.reportOAuthFailure(error: authError, flow: flow)
-                await MainActor.run {
-                    self.isLoginInProgress = false
-                    completion(.failure(authError))
-                    self.setIsLoading(false)
-                    self.setWebLoading(false)
-                }
-                return
-            }
-            
-            logger.info("Hosted login token exchange succeeded")
-            
-            do {
-                logger.info("Going to load user data")
-                let meResult = try await self.api.me(
-                    accessToken: data.access_token,
-                    refreshToken: data.refresh_token
-                )
-                let hydratedTokens = meResult.refreshedTokens ?? data
-                let user = meResult.user
-                
-                guard let user = user else {
-                    let authError = FronteggError.authError(.failedToLoadUserData("User data is nil"))
-                    self.clearMatchedPendingOAuthFlowIfNeeded(
-                        oauthState: oauthState,
-                        codeVerifier: codeVerifier,
-                        matchedPendingOAuthState: matchedPendingOAuthState,
-                        managePendingOAuthFlow: completePendingFlowOnSuccess
-                    )
-                    self.reportOAuthFailure(error: authError, flow: flow)
-                    await MainActor.run {
-                        self.isLoginInProgress = false
-                        completion(.failure(authError))
-                        self.setIsLoading(false)
-                        self.setWebLoading(false)
-                    }
-                    return
-                }
-                
-                await setCredentials(
-                    accessToken: hydratedTokens.access_token,
-                    refreshToken: hydratedTokens.refresh_token,
-                    user: user
-                )
-                self.completePendingOAuthFlowIfNeeded(
-                    oauthState: oauthState,
-                    codeVerifier: codeVerifier,
-                    matchedPendingOAuthState: matchedPendingOAuthState,
-                    completePendingFlowOnSuccess: completePendingFlowOnSuccess
-                )
-                
-                // Call completion on main thread to avoid race conditions
-                // This ensures the completion handler always runs on the main thread
-                // Use the user from memory state (which may have been modified to use local tenant)
-                await MainActor.run {
-                    self.isLoginInProgress = false
-                    self.setWebLoading(false)
-                    // Return the user from memory state, which may have been modified to use local tenant
-                    if let modifiedUser = self.user {
-                        completion(.success(modifiedUser))
-                    } else {
-                        completion(.success(user))
-                    }
-                }
-            } catch {
-                logger.error("Hosted login callback failed to load user data: \(error.localizedDescription)")
-                let enableOfflineMode = (try? PlistHelper.fronteggConfig())?.enableOfflineMode ?? false
-
-                if enableOfflineMode && isConnectivityError(error) {
-                    // Token exchange succeeded — tokens are valid.
-                    // Preserve them and enter offline mode instead of discarding.
-                    let fallbackUser = self.resolveBestEffortUser(
-                        accessToken: data.access_token,
-                        includeInMemory: false
-                    )
-                    if fallbackUser != nil {
-                        self.logPreservedSessionAfterAuthoritativeUserLoadFailure(
-                            error,
-                            stage: "handleHostedLoginCallback"
-                        )
-                    }
-
-                    guard let fallbackUser else {
-                        let authError = FronteggError.authError(.failedToLoadUserData(error.localizedDescription))
-                        self.clearMatchedPendingOAuthFlowIfNeeded(
-                            oauthState: oauthState,
-                            codeVerifier: codeVerifier,
-                            matchedPendingOAuthState: matchedPendingOAuthState,
-                            managePendingOAuthFlow: completePendingFlowOnSuccess
-                        )
-                        self.reportOAuthFailure(error: authError, flow: flow)
-                        await MainActor.run {
-                            self.isLoginInProgress = false
-                            completion(.failure(authError))
-                            self.setIsLoading(false)
-                            self.setWebLoading(false)
-                        }
-                        return
-                    }
-
-                    await self.setCredentialsInternal(
-                        accessToken: data.access_token,
-                        refreshToken: data.refresh_token,
-                        user: fallbackUser,
-                        hydrationMode: .preserveCachedOrDerivedUser
-                    )
-                    let hydratedSession = await MainActor.run {
-                        (isAuthenticated: self.isAuthenticated, user: self.user)
-                    }
-                    guard hydratedSession.isAuthenticated,
-                          let resolvedUser = hydratedSession.user else {
-                        let authError = FronteggError.authError(.failedToAuthenticate)
-                        self.clearMatchedPendingOAuthFlowIfNeeded(
-                            oauthState: oauthState,
-                            codeVerifier: codeVerifier,
-                            matchedPendingOAuthState: matchedPendingOAuthState,
-                            managePendingOAuthFlow: completePendingFlowOnSuccess
-                        )
-                        self.reportOAuthFailure(error: authError, flow: flow)
-                        await MainActor.run {
-                            self.isLoginInProgress = false
-                            completion(.failure(authError))
-                            self.setIsLoading(false)
-                            self.setWebLoading(false)
-                        }
-                        return
-                    }
-                    await MainActor.run {
-                        self.setIsOfflineMode(true)
-                    }
-                    self.ensureOfflineMonitoringActive(emitInitialState: false)
-                    self.completePendingOAuthFlowIfNeeded(
-                        oauthState: oauthState,
-                        codeVerifier: codeVerifier,
-                        matchedPendingOAuthState: matchedPendingOAuthState,
-                        completePendingFlowOnSuccess: completePendingFlowOnSuccess
-                    )
-                    await MainActor.run {
-                        self.isLoginInProgress = false
-                        self.setWebLoading(false)
-                        completion(.success(resolvedUser))
-                    }
-                } else {
-                    let authError = FronteggError.authError(.failedToLoadUserData(error.localizedDescription))
-                    self.clearMatchedPendingOAuthFlowIfNeeded(
-                        oauthState: oauthState,
-                        codeVerifier: codeVerifier,
-                        matchedPendingOAuthState: matchedPendingOAuthState,
-                        managePendingOAuthFlow: completePendingFlowOnSuccess
-                    )
-                    self.reportOAuthFailure(error: authError, flow: flow)
-                    await MainActor.run {
-                        self.isLoginInProgress = false
-                        completion(.failure(authError))
-                        self.setIsLoading(false)
-                        self.setWebLoading(false)
-                    }
-                }
-                return
-            }
-
-        }
-
-    }
-    
-    
-    public func getOrRefreshAccessTokenAsync() async throws -> String? {
-        self.logger.info("Waiting for isLoading | initializing | refreshingToken indicators")
-        
-        let maxAttempts = 100 // Max waiting attempts to avoid infinite loops (20 seconds total)
-        var attempt = 0
-        while (self.isLoading || self.initializing || self.refreshingToken) && attempt < maxAttempts {
-            try await Task.sleep(nanoseconds: 200_000_000) // Sleep for 200ms
-            attempt += 1
-        }
-        
-        if attempt == maxAttempts {
-            self.logger.error("Timeout while waiting for isLoading to complete (20 seconds)")
-            throw FronteggError.authError(.failedToAuthenticate)
-        }
-        
-        self.logger.info("Checking if refresh token exists")
-        
-        let config = try? PlistHelper.fronteggConfig()
-        let enableSessionPerTenant = config?.enableSessionPerTenant ?? false
-        let enableOfflineMode = config?.enableOfflineMode ?? false
-
-        // Reload both refresh and access tokens from storage using consistent precedence
-        let artifacts = resolveStoredSessionArtifacts(enableSessionPerTenant: enableSessionPerTenant)
-
-        // Reload refresh token
-        var refreshToken = self.refreshToken
-        if refreshToken == nil, let storedRefresh = artifacts.refreshToken {
-            self.logger.info("Reloaded refresh token from keychain in getOrRefreshAccessTokenAsync")
-            await MainActor.run { setRefreshToken(storedRefresh) }
-            refreshToken = storedRefresh
-        }
-
-        // Reload access token from keychain if in-memory is nil
-        if self.accessToken == nil, let storedAccess = artifacts.accessToken {
-            self.logger.info("Reloaded access token from keychain in getOrRefreshAccessTokenAsync")
-            await MainActor.run { setAccessToken(storedAccess) }
-        }
-
-        guard let refreshToken = refreshToken else {
-            self.logger.info("No refresh token found in memory or keychain")
-            return nil
-        }
-
-        self.logger.info("Check if the access token exists and is still valid")
-
-        if let accessToken = self.accessToken {
-            do {
-                let decodedToken = try JWTHelper.decode(jwtToken: accessToken)
-                if let exp = decodedToken["exp"] as? Int {
-                    let offset = self.calculateOffset(expirationTime: exp)
-                    self.logger.info("Access token offset: \(offset)")
-                    if offset > 15 { // Ensure token has more than 15 seconds validity
-                        return accessToken
-                    }
-                }
-            } catch {
-                self.logger.error("Failed to decode JWT: \(error.localizedDescription)")
-            }
-        }
-
-        if enableOfflineMode {
-            let isNetworkAvailable = await checkNetworkPath(timeout: 300_000_000)
-            if !isNetworkAvailable {
-                self.logger.info("getOrRefreshAccessTokenAsync: NWPathMonitor reported offline, but continuing with manual refresh because the path check is advisory")
-            }
-        }
-
-        self.logger.info("Refreshing access token")
-        
-        setRefreshingToken(true)
-        defer {
-            setRefreshingToken(false)
-        }
-        
-        // Reuse config and enableSessionPerTenant from earlier in the function
-        var currentTenantId: String? = nil
-        
-        if enableSessionPerTenant {
-            // Prioritize lastActiveTenantId for per-tenant session isolation
-            if let localTenantId = credentialManager.getLastActiveTenantId() {
-                currentTenantId = localTenantId
-                self.logger.info("Using LOCAL tenant ID for refresh in getOrRefreshAccessTokenAsync: \(localTenantId)")
-            } else if let user = self.user {
-                currentTenantId = user.activeTenant.id
-                self.logger.info("No local tenant stored, using user's active tenant: \(currentTenantId!)")
-            } else if let offlineUser = credentialManager.getOfflineUser() {
-                currentTenantId = offlineUser.activeTenant.id
-                self.logger.info("No local tenant stored, using offline user's tenant: \(currentTenantId!)")
-            }
-        }
-        
-        var attempts = 0
-        while attempts < 5 {
-            do {
-                var data: AuthResponse
-                
-                if enableSessionPerTenant, let tenantId = currentTenantId {
-                    // Try tenant-specific refresh first
-                    // Include access token if available - it may be needed for tenant-specific refresh
-                    let currentAccessToken = self.accessToken
-                    do {
-                        self.logger.info("Attempting tenant-specific refresh with tenantId: \(tenantId) in getOrRefreshAccessTokenAsync")
-                        data = try await self.api.refreshToken(refreshToken: refreshToken, tenantId: tenantId, accessToken: currentAccessToken)
-                        self.logger.info("Tenant-specific refresh successful in getOrRefreshAccessTokenAsync")
-                    } catch {
-                        // Keep the fallback broad here too: the tenant endpoint can reject a refresh
-                        // while the standard endpoint still accepts it during recovery/migration.
-                        self.logger.warning("Tenant-specific refresh failed in getOrRefreshAccessTokenAsync: \(error). Falling back to standard OAuth refresh.")
-                        data = try await self.api.refreshToken(refreshToken: refreshToken, tenantId: nil)
-                        self.logger.info("Standard OAuth refresh successful (fallback) in getOrRefreshAccessTokenAsync")
-                    }
-                } else {
-                    // Standard OAuth refresh for non-per-tenant sessions
-                    data = try await self.api.refreshToken(refreshToken: refreshToken, tenantId: nil)
-                }
-                
-                await self.setCredentialsInternal(accessToken: data.access_token, refreshToken: data.refresh_token, hydrationMode: .refreshPreserveCachedUser)
-                self.logger.info("Token refreshed successfully")
-                return self.accessToken
-            } catch let error as FronteggError {
-                if case .authError(FronteggError.Authentication.failedToRefreshToken(let message)) = error {
-                    SentryHelper.logMessage(
-                        "Api: failed to refresh token, error: \(message)",
-                        level: .error,
-                        context: [
-                            "auth": [
-                                "method": "getOrRefreshAccessTokenAsync",
-                                "attempts": attempts + 1,
-                                "enableSessionPerTenant": enableSessionPerTenant
-                            ],
-                            "error": [
-                                "type": "failed_to_refresh_token",
-                                "message": message
-                            ]
-                        ]
-                    )
-                    return nil
-                }
-                self.logger.error("Failed to refresh token: \(error.localizedDescription), retrying... (\(attempts + 1) attempts)")
-                
-                SentryHelper.logError(error, context: [
-                    "auth": [
-                        "method": "getOrRefreshAccessTokenAsync",
-                        "attempts": attempts + 1,
-                        "enableSessionPerTenant": enableSessionPerTenant
-                    ],
-                    "error": [
-                        "type": "token_refresh_error"
-                    ]
-                ])
-                
-                // Note: FronteggError is a custom enum, not a URL/POSIX error.
-                // Real connectivity errors (URLError) are caught by the generic catch below.
-                attempts += 1
-                try await Task.sleep(nanoseconds: 1_000_000_000) // Sleep for 1 second before retrying
-            } catch {
-                self.logger.error("Unknown error while refreshing token: \(error.localizedDescription), retrying... (\(attempts + 1) attempts)")
-
-                SentryHelper.logError(error, context: [
-                    "auth": [
-                        "method": "getOrRefreshAccessTokenAsync",
-                        "attempts": attempts + 1,
-                        "enableSessionPerTenant": enableSessionPerTenant
-                    ],
-                    "error": [
-                        "type": "unknown_token_refresh_error"
-                    ]
-                ])
-
-                // On connectivity error while offline, return cached token instead of retrying
-                if enableOfflineMode && isConnectivityError(error) {
-                    self.logger.info("Connectivity error in getOrRefreshAccessTokenAsync, preserving cached session and returning cached token")
-                    await applyConnectivityLossState(enableOfflineMode: enableOfflineMode)
-                    self.startOfflineMonitoringAfterManualConnectivityFailure(enableOfflineMode: enableOfflineMode)
-                    return self.accessToken
-                }
-
-                attempts += 1
-                try await Task.sleep(nanoseconds: 1_000_000_000) // Sleep for 1 second before retrying
-            }
-        }
-
-        throw FronteggError.authError(.failedToAuthenticate)
-    }
-    
-    public func getOrRefreshAccessToken(_ completion: @escaping FronteggAuth.AccessTokenHandler) {
-        DispatchQueue.global(qos: .userInitiated).async {
-            Task {
-                do {
-                    let token = try await self.getOrRefreshAccessTokenAsync()
-                    DispatchQueue.main.async {
-                        completion(.success(token))
-                    }
-                } catch let error as FronteggError {
-                    DispatchQueue.main.async {
-                        completion(.failure(error))
-                    }
-                } catch {
-                    self.logger.error("Failed to get or refresh token: \(error.localizedDescription)")
-                    DispatchQueue.main.async {
-                        completion(.failure(error))
-                    }
-                }
-            }
-        }
-    }
-    func pendingOAuthState(from url: URL) -> String? {
-        CredentialManager.pendingOAuthState(from: url)
-    }
-
-    private func matchesGeneratedRedirectUri(_ url: URL) -> Bool {
-        guard
-            let expected = URLComponents(string: generateRedirectUri()),
-            let actual = URLComponents(url: url, resolvingAgainstBaseURL: false)
-        else {
-            return false
-        }
-
-        return actual.scheme == expected.scheme
-            && actual.host == expected.host
-            && actual.path == expected.path
-    }
-
-    internal func createOauthCallbackHandler(
-        _ completion: @escaping FronteggAuth.CompletionHandler,
-        allowLastCodeVerifierFallback: Bool = false,
-        redirectUriOverride: String? = nil,
-        pendingOAuthState: String? = nil,
-        flow: FronteggOAuthFlow = .login
-    ) -> ((URL?, Error?) -> Void) {
-        
-        return { [weak self] callbackUrl, error in
-            guard let self = self else { return }
-
-            func clearPendingOAuthState(_ state: String?) {
-                guard let state, !state.isEmpty else {
-                    return
-                }
-                CredentialManager.clearPendingOAuth(state: state)
-            }
-
-            func clearRelevantPendingOAuthState(callbackState: String?) {
-                clearPendingOAuthState(callbackState ?? pendingOAuthState)
-            }
-            
-            if let error {
-                clearPendingOAuthState(pendingOAuthState)
-                let fronteggError = FronteggError.authError(.other(error))
-                self.reportOAuthFailure(error: fronteggError, flow: flow)
-                completion(.failure(fronteggError))
-                return
-            }
-            
-            guard let url = callbackUrl else {
-                clearPendingOAuthState(pendingOAuthState)
-                let fronteggError = FronteggError.authError(.unknown)
-                self.reportOAuthFailure(error: fronteggError, flow: flow)
-                completion(.failure(fronteggError))
-                return
-            }
-            
-            let parsedQueryItems = getQueryItems(url.absoluteString)
-            guard let queryItems = parsedQueryItems else {
-                clearPendingOAuthState(pendingOAuthState)
-                let fronteggError = FronteggError.authError(.failedToExtractCode)
-                self.reportOAuthFailure(error: fronteggError, flow: flow)
-                completion(.failure(fronteggError))
-                return
-            }
-
-            let oauthState = queryItems["state"]
-
-            if let failureDetails = self.oauthFailureDetails(from: queryItems) {
-                clearRelevantPendingOAuthState(callbackState: oauthState)
-                self.reportOAuthFailure(details: failureDetails, flow: flow)
-                completion(.failure(failureDetails.error))
-                return
-            }
-            
-            guard let code = queryItems["code"] else {
-                let keys = Array(queryItems.keys).sorted()
-                SentryHelper.logMessage(
-                    "OAuth callback missing code (hosted)",
-                    level: .warning,
-                    context: [
-                        "oauth": [
-                            "stage": "createOauthCallbackHandler",
-                            "url": url.absoluteString,
-                            "queryKeys": keys
-                        ],
-                        "error": [
-                            "type": "oauth_missing_code"
-                        ]
-                    ]
-                )
-                // If this is a verification callback and there's no code, the verification might have succeeded
-                // but the redirect didn't include the code. In this case, we should inform the user
-                // that verification succeeded but they need to try logging in again.
-                clearRelevantPendingOAuthState(callbackState: oauthState)
-                let fronteggError = FronteggError.authError(.failedToExtractCode)
-                self.reportOAuthFailure(error: fronteggError, flow: flow)
-                completion(.failure(fronteggError))
-                return
-            }
-            
-            let resolution = CredentialManager.resolveCodeVerifier(
-                for: oauthState,
-                allowFallback: allowLastCodeVerifierFallback
-            )
-            guard let codeVerifier = resolution.verifier else {
-                let fronteggError = FronteggError.authError(
-                    self.oauthCodeVerifierError(for: oauthState, resolution: resolution)
-                )
-                self.reportOAuthFailure(error: fronteggError, flow: flow)
-                completion(.failure(fronteggError))
-                return
-            }
-            if resolution.source == .lastGeneratedFallback {
-                self.logger.warning(
-                    "Using last generated code verifier fallback for OAuth callback (state present: \(oauthState != nil))"
-                )
-            }
-
-            self.handleHostedLoginCallback(
-                code,
-                codeVerifier,
-                oauthState: oauthState,
-                redirectUri: redirectUriOverride,
-                flow: flow,
-                completePendingFlowOnSuccess: true,
-                matchedPendingOAuthState: resolution.source == .stateMatch,
-                completion: completion
-            )
-        }
-        
-    }
-    public typealias CompletionHandler = (Result<User, FronteggError>) -> Void
-    
-    public typealias AccessTokenHandler = (Result<String?, Error>) -> Void
-    
-    public typealias LogoutHandler = (Result<Bool, FronteggError>) -> Void
-    
-    public typealias ConditionCompletionHandler = (_ error: FronteggError?) -> Void
-    
-    public func login(_ _completion: FronteggAuth.CompletionHandler? = nil, loginHint: String? = nil) {
-        
-        if(self.embeddedMode){
-            self.embeddedLogin(_completion, loginHint: loginHint)
-            return
-        }
-        
-        let completion = _completion ?? { res in
-            
-        }
-        
-        let (authorizeUrl, codeVerifier) = AuthorizeUrlGenerator.shared.generate(loginHint: loginHint)
-        let oauthCallback = createOauthCallbackHandler(
-            completion,
-            pendingOAuthState: pendingOAuthState(from: authorizeUrl),
-            flow: .login
-        )
-        
-        
-        WebAuthenticator.shared.start(authorizeUrl, completionHandler: oauthCallback)
-    }
-    
-    
-    func saveWebCredentials(domain: String, email: String, password: String, completion: @escaping (Bool, Error?) -> Void) {
-        let domainString = domain
-        let account = email
-        
-        SecAddSharedWebCredential(domainString as CFString, account as CFString, password as CFString) { error in
-            if let error = error {
-                self.logger.error("Failed to save shared web credentials: \(error.localizedDescription)")
-                completion(false, error)
-            } else {
-                self.logger.info("Shared web credentials saved successfully")
-                completion(true, nil)
-            }
-        }
-    }
-    
-    
-    
-    public func loginWithPopup(window: UIWindow?, ephemeralSession: Bool? = true, loginHint: String? = nil, loginAction: String? = nil, _completion: FronteggAuth.CompletionHandler? = nil) {
-        
-        let completion = _completion ?? { res in
-            
-        }
-        
-        let (authorizeUrl, codeVerifier) = AuthorizeUrlGenerator.shared.generate(loginHint: loginHint, loginAction: loginAction)
-        let oauthCallback = createOauthCallbackHandler(
-            completion,
-            allowLastCodeVerifierFallback: loginAction != nil,
-            pendingOAuthState: pendingOAuthState(from: authorizeUrl),
-            flow: .login
-        )
-        WebAuthenticator.shared.start(authorizeUrl, ephemeralSession: ephemeralSession ?? true, window:window,  completionHandler: oauthCallback)
-    }
-    
-    private func completeSocialLoginFailure(
-        _ error: FronteggError,
-        errorCode: String? = nil,
-        errorDescription: String? = nil,
-        completion: @escaping FronteggAuth.CompletionHandler
-    ) {
-        DispatchQueue.main.async {
-            self.activeEmbeddedOAuthFlow = .login
-            self.reportOAuthFailure(
-                error: error,
-                flow: .socialLogin,
-                errorCode: errorCode,
-                errorDescription: errorDescription
-            )
-            completion(.failure(error))
-        }
-    }
-
-    private func completeSocialLoginFailure(
-        _ details: OAuthFailureDetails,
-        completion: @escaping FronteggAuth.CompletionHandler
-    ) {
-        DispatchQueue.main.async {
-            self.activeEmbeddedOAuthFlow = .login
-            self.reportOAuthFailure(details: details, flow: .socialLogin)
-            completion(.failure(details.error))
-        }
-    }
-
-    internal func handleSocialLoginOAuthCallback(
-        providerString: String,
-        callbackURL: URL?,
-        error: Error?,
-        completion: @escaping FronteggAuth.CompletionHandler
-    ) {
-        if let error {
-            self.logger.error("OAuth error: \(String(describing: error))")
-            self.completeSocialLoginFailure(
-                FronteggError.authError(.other(error)),
-                completion: completion
-            )
-            return
-        }
-
-        guard let callbackURL else {
-            self.logger.info("OAuth callback invoked with nil URL and no error")
-            self.completeSocialLoginFailure(
-                FronteggError.authError(.unknown),
-                completion: completion
-            )
-            return
-        }
-
-        self.logger.debug("OAuth callback URL: \(callbackURL.absoluteString)")
-
-        if let queryItems = getQueryItems(callbackURL.absoluteString),
-           let failureDetails = self.oauthFailureDetails(from: queryItems) {
-            self.completeSocialLoginFailure(failureDetails, completion: completion)
-            return
-        }
-
-        DispatchQueue.main.async { [weak self] in
-            guard let self else {
-                return
-            }
-
-            guard let finalURL = self.handleSocialLoginCallback(callbackURL) else {
-                SentryHelper.logMessage(
-                    "Social login callback could not be parsed (hosted)",
-                    level: .warning,
-                    context: [
-                        "social_login": [
-                            "provider": providerString,
-                            "callbackUrl": callbackURL.absoluteString,
-                            "baseUrl": FronteggAuth.shared.baseUrl
-                        ],
-                        "error": [
-                            "type": "social_login_callback_unhandled"
-                        ]
-                    ]
-                )
-                self.completeSocialLoginFailure(
-                    FronteggError.authError(.failedToExtractCode),
-                    completion: completion
-                )
-                return
-            }
-            self.loadInWebView(finalURL)
-        }
-    }
-
-    /// Starts a social login flow.
-    ///
-    /// - Parameters:
-    ///   - providerString: The social provider raw value (e.g., "google", "facebook", "apple").
-    ///   - action: The social login action to perform (default: `.login`).
-    ///   - completion: Optional completion handler. A no-op is used if nil.
-    public func handleSocialLogin(
-        providerString: String,
-        custom:Bool,
-        action: SocialLoginAction = .login,
-        completion: FronteggAuth.CompletionHandler? = nil
-    ) {
-        FronteggRuntime.testingLog(
-            "E2E handleSocialLogin start provider=\(providerString) custom=\(custom) action=\(action.rawValue)"
-        )
-        let done = completion ?? { _ in }
-        
-        // Special-case Apple to keep branching explicit and fast.
-        if providerString == "apple" {
-            loginWithApple(done)
-            return
-        }
-        
-        let oauthCallback: (URL?, Error?) -> Void = { [weak self] callbackURL, error in
-            guard let self else { return }
-            self.handleSocialLoginOAuthCallback(
-                providerString: providerString,
-                callbackURL: callbackURL,
-                error: error,
-                completion: done
-            )
-        }
-        
-        Task { [weak self] in
-            guard let self else { return }
-            
-            let generatedAuthUrl: URL? = if(custom){
-                try? await SocialLoginUrlGenerator.shared
-                    .authorizeURL(forCustomProvider: providerString, action: action)
-            } else if let provider = SocialLoginProvider(rawValue: providerString) {
-                try? await SocialLoginUrlGenerator.shared
-                    .authorizeURL(for: provider, action: action)
-            }else {
-                nil
-            }
-            FronteggRuntime.testingLog(
-                "E2E handleSocialLogin generatedAuthUrl provider=\(providerString) url=\(generatedAuthUrl?.absoluteString ?? "nil")"
-            )
-
-            // Check if we need to use legacy flow
-            if generatedAuthUrl == nil && !custom {
-                if let provider = SocialLoginProvider(rawValue: providerString),
-                   let legacyUrl = try? await SocialLoginUrlGenerator.shared.legacyAuthorizeURL(for: provider, action: action) {
-                    logger.debug("Using legacy social login flow for provider: \(providerString)")
-                    self.loginWithSocialLogin(socialLoginUrl: legacyUrl, done)
-                    return
-                }
-            }
-            
-            guard let authURL = generatedAuthUrl else {
-                self.logger.error("Failed to generate auth URL for \(providerString)")
-                self.completeSocialLoginFailure(
-                    FronteggError.authError(.unknown),
-                    completion: done
-                )
-                return
-            }
-            
-            self.logger.debug("Auth URL: \(authURL.absoluteString)")
-           
-             let window: UIWindow? = await MainActor.run {
-                return self.getRootVC()?.view.window
-            }
-            
-            let useEphemeral = false
-            
-            await MainActor.run {
-                FronteggRuntime.testingLog(
-                    "E2E handleSocialLogin starting WebAuthenticator url=\(authURL.absoluteString)"
-                )
-                WebAuthenticator.shared.start(
-                    authURL,
-                    ephemeralSession: useEphemeral,
-                    window: window,
-                    completionHandler: oauthCallback
-                )
-            }
-        }
-    }
-    
-    private func loadInWebView(_ url: URL) {
-        guard let webView = webview else { return }
-        let request = URLRequest(url: url, cachePolicy: .reloadRevalidatingCacheData)
-        webView.load(request)
-    }
-    
-    public func directLoginAction(
-        window: UIWindow?,
-        type: String,
-        data: String,
-        ephemeralSession: Bool? = true,
-        _completion: FronteggAuth.CompletionHandler? = nil,
-        additionalQueryParams: [String: Any]? = nil,
-        remainCodeVerifier: Bool = false,
-        action: SocialLoginAction = SocialLoginAction.login
-    ) {
-        
-        let completion = _completion ?? { res in
-            
-        }
-        
-        if(type == "social-login" && data == "apple") {
-            self.loginWithApple(completion)
-            return
-        }
-        
-        
-        var directLogin = [
-            "type": type,
-            "data": data,
-            
-        ] as [String : Any]
-        
-        if let queryParams = additionalQueryParams {
-            directLogin["additionalQueryParams"] = queryParams
-        }
-        
-        var generatedUrl: (URL, String)
-        if let jsonData = try? JSONSerialization.data(withJSONObject: directLogin, options: []) {
-            let jsonString = jsonData.base64EncodedString()
-            generatedUrl = AuthorizeUrlGenerator.shared.generate(loginAction: jsonString, remainCodeVerifier: remainCodeVerifier)
-        } else {
-            generatedUrl = AuthorizeUrlGenerator.shared.generate(remainCodeVerifier: remainCodeVerifier)
-        }
-        
-        let (authorizeUrl, codeVerifier) = generatedUrl
-        let oauthFlow: FronteggOAuthFlow
-        switch type {
-        case "social-login", "custom-social-login":
-            oauthFlow = .socialLogin
-        case "direct" where data.contains("/user/sso/") || data.contains("appleid.apple.com"):
-            oauthFlow = .socialLogin
-        default:
-            oauthFlow = .login
-        }
-        let oauthCallback = createOauthCallbackHandler(
-            completion,
-            allowLastCodeVerifierFallback: true,
-            pendingOAuthState: pendingOAuthState(from: authorizeUrl),
-            flow: oauthFlow
-        )
-        
-        WebAuthenticator.shared.start(authorizeUrl, ephemeralSession: ephemeralSession ?? true, window: window ?? getRootVC()?.view.window, completionHandler: oauthCallback)
-    }
-    
-    
-    internal func getRootVC(_ useAppRootVC: Bool = false) -> UIViewController? {
-        
-        
-        if let appDelegate = UIApplication.shared.delegate,
-           let window = appDelegate.window,
-           let rootVC = window?.rootViewController {
-            
-            if(useAppRootVC){
-                return rootVC
-            }else {
-                if let presented = rootVC.presentedViewController {
-                    return presented
-                }else {
-                    return rootVC
-                }
-            }
-        }
-        
-        if let rootVC = UIWindow.key?.rootViewController {
-            return rootVC
-        }
-        if let lastWindow = UIApplication.shared.windows.last,
-           let rootVC = lastWindow.rootViewController {
-            return rootVC
-        }
-        
-        
-        return nil
-    }
-    
-    
-    
-    internal func loginWithApple(_ _completion: @escaping FronteggAuth.CompletionHandler)  {
-        let completion = handleMfaRequired(_completion)
-        DispatchQueue.global(qos: .userInitiated).async {
-            Task {
-                do{
-                    let config = try PlistHelper.fronteggConfig()
-                    let socialConfig =  try await self.api.getSocialLoginConfig()
-                    
-                    if let appleConfig = socialConfig.apple, appleConfig.active {
-                        if #available(iOS 15.0, *), appleConfig.customised, !config.useAsWebAuthenticationForAppleLogin {
-                            await AppleAuthenticator.shared.start(completionHandler: completion)
-                        }else {
-                            let generatedAuth = try await self.generateAppleAuthorizeUrl(config: appleConfig)
-                            let oauthCallback = self.createOauthCallbackHandler(
-                                completion,
-                                allowLastCodeVerifierFallback: true,
-                                pendingOAuthState: generatedAuth.pendingOAuthState,
-                                flow: .apple
-                            )
-                            await WebAuthenticator.shared.start(generatedAuth.url, ephemeralSession: true, completionHandler: oauthCallback)
-                        }
-                    } else {
-                        throw FronteggError.configError(.socialLoginMissing("Apple"))
-                    }
-                } catch {
-                    if error is FronteggError {
-                        completion(.failure(error as! FronteggError))
-                    }else {
-                        self.logger.error(error.localizedDescription)
-                        completion(.failure(FronteggError.authError(.unknown)))
-                    }
-                    
-                }
-            }
-            
-        }
-    }
-    
-    
-    private struct OAuth2SessionStateContext {
-        let encodedState: Data
-        let pendingOAuthState: String?
-    }
-
-    private func createOauth2SessionState() async throws -> OAuth2SessionStateContext {
-        let (url, codeVerifier)  = AuthorizeUrlGenerator.shared.generate()
-        let pendingOAuthState = pendingOAuthState(from: url)
-        
-        let (_, authorizeResponse) = try await FronteggAuth.shared.api.getRequest(path: url.absoluteString, accessToken: nil, additionalHeaders: ["Accept":"text/html"])
-        
-        guard let authorizeResponseUrl = authorizeResponse.url,
-              let authorizeComponent = URLComponents(string: authorizeResponseUrl.absoluteString),
-              let sessionState = authorizeComponent.queryItems?.first(where: { q in
-                  q.name == "state"
-              })?.value else {
-            throw FronteggError.authError(.failedToAuthenticate)
-        }
-        
-        let redirectUri = generateRedirectUri()
-        
-        let oauthStateDic = [
-            "FRONTEGG_OAUTH_REDIRECT_AFTER_LOGIN": redirectUri,
-            "FRONTEGG_OAUTH_STATE_AFTER_LOGIN": sessionState,
-        ]
-        
-        guard let oauthStateJson = try? JSONSerialization.data(withJSONObject: oauthStateDic, options: .withoutEscapingSlashes) else {
-            throw FronteggError.authError(.failedToAuthenticate)
-        }
-
-        return OAuth2SessionStateContext(
-            encodedState: oauthStateJson,
-            pendingOAuthState: pendingOAuthState
-        )
-    }
-    func generateAppleAuthorizeUrl(config: SocialLoginOption) async throws -> (url: URL, pendingOAuthState: String?) {
-        let sessionState = try await createOauth2SessionState()
-        
-        let scope = ["openid", "name", "email"] + config.additionalScopes
-        let appId = FronteggAuth.shared.applicationId ?? ""
-        
-        let stateDict = [
-            "oauthState": sessionState.encodedState.base64EncodedString(),
-            "appId": appId,
-            "provider": "apple",
-            "action": "login",
-        ]
-        
-        guard let stateJson = try? JSONSerialization.data(withJSONObject: stateDict, options: .withoutEscapingSlashes),
-              let state = String(data: stateJson, encoding: .utf8) else {
-            throw FronteggError.authError(.failedToAuthenticate)
-        }
-        
-        var urlComponent = URLComponents(string: "https://appleid.apple.com/auth/authorize")!
-        urlComponent.queryItems = [
-            URLQueryItem(name: "response_type", value: "code"),
-            URLQueryItem(name: "response_mode", value: "form_post"),
-            URLQueryItem(name: "redirect_uri", value: config.backendRedirectUrl),
-            URLQueryItem(name: "scope", value: scope.joined(separator: " ")),
-            URLQueryItem(name: "state", value: state),
-            URLQueryItem(name: "client_id", value: config.clientId)
-        ]
-        
-        let finalUrl = urlComponent.url!
-        
-        return (finalUrl, sessionState.pendingOAuthState)
-    }
-    
-    func loginWithSocialLogin(socialLoginUrl: String, _ _completion: FronteggAuth.CompletionHandler? = nil) {
-        let completion = _completion ?? self.loginCompletion ?? { res in
-            
-            
-        }
-        
-        // Log social login initiation
-        let isMicrosoft = socialLoginUrl.contains("microsoft") || socialLoginUrl.contains("login.microsoftonline.com")
-        // Use shared session (non-ephemeral) for all providers to show saved accounts
-        let useEphemeral = false
-        
-        SentryHelper.addBreadcrumb(
-            "Social login initiated (loginWithSocialLogin)",
-            category: "social_login",
-            level: .info,
-            data: [
-                "socialLoginUrl": socialLoginUrl,
-                "isMicrosoft": isMicrosoft,
-                "useEphemeral": useEphemeral,
-                "embeddedMode": self.embeddedMode,
-                "baseUrl": self.baseUrl
-            ]
-        )
-        
-        let directLogin: [String: Any] = [
-            "type": "direct",
-            "data": socialLoginUrl,
-        ]
-        var generatedUrl: (URL, String)
-        if let jsonData = try? JSONSerialization.data(withJSONObject: directLogin, options: []) {
-            let jsonString = jsonData.base64EncodedString()
-            generatedUrl = AuthorizeUrlGenerator.shared.generate(loginAction: jsonString, remainCodeVerifier: true)
-        } else {
-            generatedUrl = AuthorizeUrlGenerator.shared.generate(remainCodeVerifier: true)
-        }
-        
-        let (authorizeUrl, codeVerifier) = generatedUrl
-        let oauthCallback = createOauthCallbackHandler(
-            completion,
-            allowLastCodeVerifierFallback: true,
-            pendingOAuthState: pendingOAuthState(from: authorizeUrl),
-            flow: .socialLogin
-        )
-        
-        logger.info("🔵 [Social Login] Starting social login flow")
-        logger.info("🔵 [Social Login] Authorize URL: \(authorizeUrl.absoluteString)")
-        logger.info("🔵 [Social Login] Use ephemeral session: \(useEphemeral)")
-        
-       let window: UIWindow?
-        if Thread.isMainThread {
-            window = getRootVC()?.view.window
-        } else {
-            var mainWindow: UIWindow?
-            DispatchQueue.main.sync {
-                mainWindow = getRootVC()?.view.window
-            }
-            window = mainWindow
-        }
-        
-         if Thread.isMainThread {
-            WebAuthenticator.shared.start(authorizeUrl, ephemeralSession: useEphemeral, window: window, completionHandler: oauthCallback)
-        } else {
-            DispatchQueue.main.async {
-                WebAuthenticator.shared.start(authorizeUrl, ephemeralSession: useEphemeral, window: window, completionHandler: oauthCallback)
-            }
-        }
-    }
-    
-    
-    public func loginWithSSO(email: String, _ _completion: FronteggAuth.CompletionHandler? = nil) {
-        let completion = _completion ?? self.loginCompletion ?? { res in
-            
-        }
-        let (authorizeUrl, codeVerifier) = AuthorizeUrlGenerator.shared.generate(loginHint: email, remainCodeVerifier: true)
-        let oauthCallback = createOauthCallbackHandler(
-            completion,
-            allowLastCodeVerifierFallback: true,
-            pendingOAuthState: pendingOAuthState(from: authorizeUrl),
-            flow: .sso
-        )
-        
-        WebAuthenticator.shared.start(authorizeUrl, ephemeralSession: true, window: getRootVC()?.view.window, completionHandler: oauthCallback)
-    }
-    
-    public func loginWithCustomSSO(ssoUrl: String, _ _completion: FronteggAuth.CompletionHandler? = nil) {
-        let completion = _completion ?? self.loginCompletion ?? { res in
-            
-        }
-        
-        let directLogin: [String: Any] = [
-            "type": "direct",
-            "data": ssoUrl,
-        ]
-        var generatedUrl: (URL, String)
-        if let jsonData = try? JSONSerialization.data(withJSONObject: directLogin, options: []) {
-            let jsonString = jsonData.base64EncodedString()
-            generatedUrl = AuthorizeUrlGenerator.shared.generate(loginAction: jsonString, remainCodeVerifier: true)
-        } else {
-            generatedUrl = AuthorizeUrlGenerator.shared.generate(remainCodeVerifier: true)
-        }
-        
-        let (authorizeUrl, codeVerifier) = generatedUrl
-        let oauthCallback = createOauthCallbackHandler(
-            completion,
-            allowLastCodeVerifierFallback: true,
-            pendingOAuthState: pendingOAuthState(from: authorizeUrl),
-            flow: .customSSO
-        )
-        FronteggRuntime.testingLog("loginWithCustomSSO authorizeUrl: \(authorizeUrl.absoluteString)")
-
-        WebAuthenticator.shared.start(
-            authorizeUrl,
-            ephemeralSession: true,
-            window: getRootVC()?.view.window,
-            completionHandler: oauthCallback
-        )
-    }
-    
-    public func embeddedLogin(_ _completion: FronteggAuth.CompletionHandler? = nil, loginHint: String?) {
-        
-        if let rootVC = self.getRootVC() {
-            FronteggRuntime.testingLog(
-                "E2E embeddedLogin rootVC=\(type(of: rootVC)) presented=\(String(describing: rootVC.presentedViewController)) embeddedMode=\(self.embeddedMode)"
-            )
-            self.loginHint = loginHint
-            if self.pendingAppLink == nil {
-                self.activeEmbeddedOAuthFlow = .login
-            }
-            if self.loginCompletion != nil {
-                logger.info("Login request ignored, Embedded login already in progress.")
-                return
-            }
-            self.loginCompletion = { result in
-                _completion?(result)
-                self.loginCompletion = nil
-            }
-            let loginModal = EmbeddedLoginModal(parentVC: rootVC)
-            let hostingController = UIHostingController(rootView: loginModal)
-            hostingController.modalPresentationStyle = .fullScreen
-            
-            if(rootVC.presentedViewController?.classForCoder == hostingController.classForCoder){
-                rootVC.presentedViewController?.dismiss(animated: false)
-            }
-            
-            rootVC.present(hostingController, animated: false, completion: nil)
-            FronteggRuntime.testingLog("E2E embeddedLogin present called")
-
-        } else {
-            logger.critical(FronteggError.authError(.couldNotFindRootViewController).localizedDescription)
-            exit(500)
-        }
-    }
-    public func handleOpenUrl(_ url: URL, _ useAppRootVC: Bool = false, internalHandleUrl:Bool = false) -> Bool {
-        logger.info("🔵 [handleOpenUrl] Received URL: \(url.absoluteString)")
-        logger.info("🔵 [handleOpenUrl] Base URL: \(self.baseUrl)")
-        logger.info("🔵 [handleOpenUrl] URL has prefix baseUrl: \(url.absoluteString.hasPrefix(self.baseUrl))")
-        logger.info("🔵 [handleOpenUrl] internalHandleUrl: \(internalHandleUrl)")
-        let matchesGeneratedRedirectCallback = matchesGeneratedRedirectUri(url)
-        let parsedQueryItems = getQueryItems(url.absoluteString)
-
-        // Log app redirect handling
-        SentryHelper.addBreadcrumb(
-            "App redirect received (handleOpenUrl)",
-            category: "app_redirect",
-            level: .info,
-            data: [
-                "url": url.absoluteString,
-                "scheme": url.scheme ?? "nil",
-                "host": url.host ?? "nil",
-                "path": url.path,
-                "query": url.query ?? "nil",
-                "baseUrl": self.baseUrl,
-                "matchesBaseUrl": url.absoluteString.hasPrefix(self.baseUrl),
-                "matchesGeneratedRedirectUri": matchesGeneratedRedirectCallback,
-                "internalHandleUrl": internalHandleUrl,
-                "embeddedMode": self.embeddedMode
-            ]
-        )
-
-        let redirectCallbackFailureDetails: OAuthFailureDetails? = {
-            guard let parsedQueryItems else { return nil }
-            if let failureDetails = self.oauthFailureDetails(from: parsedQueryItems) {
-                return failureDetails
-            }
-
-            let rawError = parsedQueryItems["error"]?.trimmingCharacters(in: .whitespacesAndNewlines)
-            let rawDescription = parsedQueryItems["error_description"]?.trimmingCharacters(in: .whitespacesAndNewlines)
-            guard !(rawError?.isEmpty ?? true) || !(rawDescription?.isEmpty ?? true) else {
-                return nil
-            }
-
-            return self.oauthFailureDetails(
-                errorCode: parsedQueryItems["error"],
-                errorDescription: parsedQueryItems["error_description"],
-                fallbackError: FronteggError.authError(.failedToExtractCode)
-            )
-        }()
-
-        if matchesGeneratedRedirectCallback,
-           let failureDetails = redirectCallbackFailureDetails {
-            logger.info("✅ [handleOpenUrl] Detected generated redirect URI OAuth error callback")
-            self.reportOAuthFailure(
-                details: failureDetails,
-                flow: self.activeEmbeddedOAuthFlow
-            )
-
-            if let webView = self.webview {
-                let (newUrl, _) = AuthorizeUrlGenerator().generate(remainCodeVerifier: true)
-                self.setWebLoading(false)
-                DispatchQueue.main.asyncAfter(deadline: .now() + 0.15) {
-                    webView.load(URLRequest(url: newUrl, cachePolicy: .reloadRevalidatingCacheData))
-                }
-            }
-
-            self.activeEmbeddedOAuthFlow = .login
-            setAppLink(false)
-            return true
-        }
-        
-        if(!url.absoluteString.hasPrefix(self.baseUrl) && !internalHandleUrl && !matchesGeneratedRedirectCallback){
-            logger.warning("⚠️ [handleOpenUrl] URL doesn't match baseUrl and internalHandleUrl is false, returning false")
-            SentryHelper.logMessage(
-                "App redirect URL rejected - doesn't match baseUrl",
-                level: .warning,
-                context: [
-                    "app_redirect": [
-                        "url": url.absoluteString,
-                        "baseUrl": self.baseUrl,
-                        "internalHandleUrl": internalHandleUrl,
-                        "matchesGeneratedRedirectUri": matchesGeneratedRedirectCallback,
-                        "scheme": url.scheme ?? "nil",
-                        "host": url.host ?? "nil"
-                    ],
-                    "error": [
-                        "type": "redirect_url_mismatch"
-                    ]
-                ]
-            )
-            setAppLink(false)
-            return false
-        }
-        
-        if url.path.contains("/postlogin/verify") {
-            logger.info("✅ [handleOpenUrl] Detected /postlogin/verify URL, processing verification")
-            SentryHelper.addBreadcrumb(
-                "Processing /postlogin/verify URL",
-                category: "app_redirect",
-                level: .info,
-                data: [
-                    "url": url.absoluteString,
-                    "hasToken": url.query?.contains("token") ?? false
-                ]
-            )
-            var verificationUrl = url
-            if var urlComponents = URLComponents(url: url, resolvingAgainstBaseURL: false) {
-                var queryItems = urlComponents.queryItems ?? []
-                
-                let redirectUri = generateRedirectUri()
-                if !queryItems.contains(where: { $0.name == "redirect_uri" }) {
-                    queryItems.append(URLQueryItem(name: "redirect_uri", value: redirectUri))
-                }
-                
-                if let codeVerifier = CredentialManager.getCodeVerifier() {
-                     if !queryItems.contains(where: { $0.name == "code_verifier_pkce" }) {
-                        queryItems.append(URLQueryItem(name: "code_verifier_pkce", value: codeVerifier))
-                        logger.info("Added code_verifier_pkce to verification URL")
-                    }
-                } else {
-                    logger.warning("No code verifier found for verification URL - this may cause verification to fail")
-                }
-                
-                urlComponents.queryItems = queryItems
-                if let updatedUrl = urlComponents.url {
-                    verificationUrl = updatedUrl
-                }
-            }
-            
-            let completion: FronteggAuth.CompletionHandler
-            if let existingCompletion = self.loginCompletion {
-                completion = existingCompletion
-            } else {
-                completion = { [weak self] result in
-                    guard let self = self else { return }
-                    switch result {
-                    case .success(let user):
-                        self.logger.info("✅ Email verification completed successfully. User logged in: \(user.email)")
-                        // Login is complete, no need to do anything else
-                    case .failure(let error):
-                        self.logger.error("❌ Email verification failed: \(error.localizedDescription)")
-                    }
-                }
-            }
-            
-            let oauthCallback = createOauthCallbackHandler(
-                completion,
-                allowLastCodeVerifierFallback: true,
-                pendingOAuthState: pendingOAuthState(from: verificationUrl),
-                flow: .verification
-            )
-            
-            var callbackReceived = false
-            var sessionRef: ASWebAuthenticationSession? = nil
-            
-            let wrappedCallback: (URL?, Error?) -> Void = { [weak self] callbackUrl, error in
-                guard let self = self else { return }
-                callbackReceived = true
-                
-                // Cancel the session immediately to prevent showing localhost
-                if let session = sessionRef {
-                    session.cancel()
-                    sessionRef = nil
-                }
-                
-                if let url = callbackUrl, let host = url.host, (host.contains("localhost") || host.contains("127.0.0.1")) {
-                    self.logger.warning("⚠️ Detected localhost redirect in verification callback, retrying social login immediately")
-                    // Retry immediately without delay
-                    if let urlComponents = URLComponents(url: verificationUrl, resolvingAgainstBaseURL: false),
-                       let queryItems = urlComponents.queryItems,
-                       let type = queryItems.first(where: { $0.name == "type" })?.value {
-                        self.handleSocialLogin(providerString: type, custom: false, action: .login) { result in
-                            switch result {
-                            case .success(let user):
-                                completion(.success(user))
-                            case .failure(let error):
-                                completion(.failure(error))
-                            }
-                        }
-                    } else {
-                        completion(.failure(FronteggError.authError(.unknown)))
-                    }
-                    return
-                }
-                
-                
-                oauthCallback(callbackUrl, error)
-            }
-            
-            // Reduce timeout to 3 seconds to minimize localhost visibility
-            // If verification takes longer, it likely redirected to localhost
-            DispatchQueue.main.asyncAfter(deadline: .now() + 3.0) { [weak self] in
-                guard let self = self, !callbackReceived else { return }
-                
-                // Cancel the session to stop showing localhost
-                if let session = sessionRef {
-                    session.cancel()
-                    sessionRef = nil
-                }
-                
-                self.logger.info("🔄 Verification timeout - retrying social login to avoid localhost redirect")
-                if let urlComponents = URLComponents(url: verificationUrl, resolvingAgainstBaseURL: false),
-                   let queryItems = urlComponents.queryItems,
-                   let type = queryItems.first(where: { $0.name == "type" })?.value {
-                    self.handleSocialLogin(providerString: type, custom: false, action: .login) { result in
-                        switch result {
-                        case .success(let user):
-                            self.logger.info("✅ Login completed successfully after verification timeout")
-                            completion(.success(user))
-                        case .failure(let error):
-                            self.logger.error("❌ Login failed after verification timeout: \(error.localizedDescription)")
-                            completion(.failure(error))
-                        }
-                    }
-                } else {
-                    completion(.failure(FronteggError.authError(.unknown)))
-                }
-            }
-            
-            let window: UIWindow?
-            if Thread.isMainThread {
-                window = getRootVC(useAppRootVC)?.view.window
-            } else {
-                var mainWindow: UIWindow?
-                DispatchQueue.main.sync {
-                    mainWindow = getRootVC(useAppRootVC)?.view.window
-                }
-                window = mainWindow
-            }
-            
-            if Thread.isMainThread {
-                WebAuthenticator.shared.start(verificationUrl, ephemeralSession: false, window: window, completionHandler: wrappedCallback)
-                sessionRef = WebAuthenticator.shared.session
-            } else {
-                DispatchQueue.main.async {
-                    WebAuthenticator.shared.start(verificationUrl, ephemeralSession: false, window: window, completionHandler: wrappedCallback)
-                    sessionRef = WebAuthenticator.shared.session
-                }
-            }
-            return true
-        }
-        
-        guard let rootVC = self.getRootVC(useAppRootVC) else {
-            self.logger.error(FronteggError.authError(.couldNotFindRootViewController).localizedDescription)
-            return false;
-        }
-        
-        if let socialLoginUrl = handleSocialLoginCallback(url){
-            self.activeEmbeddedOAuthFlow = .socialLogin
-            if let webView = self.webview {
-                let request = URLRequest(url: socialLoginUrl, cachePolicy: .reloadRevalidatingCacheData)
-                webView.load(request)
-                return true
-            }else {
-                self.pendingAppLink = socialLoginUrl
-            }
-        }else {
-            self.activeEmbeddedOAuthFlow = .login
-            self.pendingAppLink = url
-        }
-        setWebLoading(true)
-        
-        // Cancel any active ASWebAuthenticationSession before presenting EmbeddedLoginModal
-        // This prevents the magic link deep link from opening Internal WebView on top of Custom Tab
-        // which would break the session context
-        if let activeSession = WebAuthenticator.shared.session {
-            activeSession.cancel()
-            WebAuthenticator.shared.session = nil
-        }
-        
-        let loginModal = EmbeddedLoginModal(parentVC: rootVC)
-        let hostingController = UIHostingController(rootView: loginModal)
-        hostingController.modalPresentationStyle = .fullScreen
-        
-        let presented = rootVC.presentedViewController
-        if presented is UIHostingController<EmbeddedLoginModal> {
-            rootVC.presentedViewController?.dismiss(animated: false)
-        }
-        rootVC.present(hostingController, animated: false, completion: nil)
-        
-        return true
-    }
-    
-    public func  switchTenant(tenantId:String,_ completion: FronteggAuth.CompletionHandler? = nil) {
-        
-        self.logger.info("Switching tenant to: \(tenantId)")
-        if let currentUser = self.user {
-            self.logger.info("Current tenant: \(currentUser.activeTenant.name) (ID: \(currentUser.activeTenant.id), tenantId: \(currentUser.activeTenant.tenantId))")
-        }
-        
-        self.setIsLoading(true)
-        
-        DispatchQueue.global(qos: .userInitiated).async {
-            Task {
-                let config = try? PlistHelper.fronteggConfig()
-                let enableSessionPerTenant = config?.enableSessionPerTenant ?? false
-                
-                
-                if enableSessionPerTenant {
-                    self.credentialManager.saveLastActiveTenantId(tenantId)
-                    self.logger.info("Saved new tenant ID (\(tenantId)) as last active tenant before switching")
-                    
-                    if let currentUser = self.user {
-                        let currentTenantId = currentUser.activeTenant.id
-                        if let currentRefreshToken = self.refreshToken,
-                           let currentAccessToken = self.accessToken {
-                            do {
-                                try self.credentialManager.saveTokenForTenant(currentRefreshToken, tenantId: currentTenantId, tokenType: .refreshToken)
-                                try self.credentialManager.saveTokenForTenant(currentAccessToken, tenantId: currentTenantId, tokenType: .accessToken)
-                                self.logger.info("Saved tokens for tenant \(currentTenantId) before switching")
-                            } catch {
-                                self.logger.warning("Failed to save tokens for tenant \(currentTenantId): \(error)")
-                            }
-                        }
-                    }
-                    if let newRefreshToken = try? self.credentialManager.getTokenForTenant(tenantId: tenantId, tokenType: .refreshToken),
-                       let newAccessToken = try? self.credentialManager.getTokenForTenant(tenantId: tenantId, tokenType: .accessToken) {
-                        // Load existing tokens for the new tenant
-                        await MainActor.run {
-                            self.setRefreshToken(newRefreshToken)
-                            self.setAccessToken(newAccessToken)
-                        }
-                        self.logger.info("Loaded existing tokens for tenant \(tenantId) from local storage")
-                        
-                        do {
-                            let data = try await self.api.refreshToken(
-                                refreshToken: newRefreshToken,
-                                tenantId: tenantId
-                            )
-                            await self.setCredentials(accessToken: data.access_token, refreshToken: data.refresh_token)
-                            
-                            if let user = self.user {
-                                self.logger.info("Tenant switch completed using existing tokens (no server-side API call). New active tenant: \(user.activeTenant.name) (ID: \(user.activeTenant.id))")
-                                await MainActor.run {
-                                    self.setIsLoading(false)
-                                }
-                    completion?(.success(user))
-                                return
-                            }
-                        } catch {
-                            self.logger.warning("Refresh with tenantId failed, trying standard OAuth refresh: \(error)")
-                            do {
-                                let data = try await self.api.refreshToken(
-                                    refreshToken: newRefreshToken,
-                                    tenantId: nil
-                                )
-                                await self.setCredentials(accessToken: data.access_token, refreshToken: data.refresh_token)
-                                
-                                if let user = self.user {
-                                    if user.activeTenant.id == tenantId {
-                                        self.logger.info("Tenant switch completed using existing tokens (standard OAuth refresh). New active tenant: \(user.activeTenant.name) (ID: \(user.activeTenant.id))")
-                                        await MainActor.run {
-                                            self.setIsLoading(false)
-                                        }
-                                        completion?(.success(user))
-                                        return
-                } else {
-                                        self.logger.warning("Standard OAuth refresh returned wrong tenant (\(user.activeTenant.id) instead of \(tenantId)), will use server-side API")
-                                    }
-                                }
-                            } catch {
-                                self.logger.warning("Both refresh methods failed for tenant \(tenantId), will create new tokens: \(error)")
-                            }
-                        }
-                    } else {
-                        self.logger.info("No existing tokens found for tenant \(tenantId) in local storage")
-                    }
-                    self.logger.info("No existing tokens found for tenant \(tenantId), creating new tokens via server-side API")
-                }
-                
-                guard let currentAccessToken = self.accessToken else {
-                    self.logger.error("No access token available for tenant switch")
-                    await MainActor.run {
-                        self.setIsLoading(false)
-                    }
-                    completion?(.failure(FronteggError.authError(.failedToSwitchTenant)))
-                    return
-                }
-                
-                self.logger.info("Calling server-side API to switch tenant to: \(tenantId) (access token length: \(currentAccessToken.count))")
-                
-                do {
-                    try await self.api.switchTenant(tenantId: tenantId, accessToken: currentAccessToken)
-                    self.logger.info("Successfully switched tenant via API to: \(tenantId)")
-                } catch {
-                    self.logger.error("Failed to switch tenant via API: \(error)")
-                    await MainActor.run {
-                        self.setIsLoading(false)
-                    }
-                    completion?(.failure(FronteggError.authError(.failedToSwitchTenant)))
-                    return
-                }
-                
-                guard let refreshToken = self.refreshToken else {
-                    self.logger.error("No refresh token available for tenant switch")
-                    await MainActor.run {
-                        self.setIsLoading(false)
-                    }
-                    completion?(.failure(FronteggError.authError(.failedToSwitchTenant)))
-                    return
-                }
-                
-                self.logger.info("Using refresh token for tenant switch (token length: \(refreshToken.count))")
-                
-                // Refresh tokens to get updated user data with new tenant
-                do {
-                    self.logger.info("Refreshing token after tenant switch to: \(tenantId)")
-                    var data: AuthResponse
-                    
-                    if enableSessionPerTenant {
-                        // After a server-side tenant switch, we MUST use standard OAuth refresh first
-                        // to get a new refresh token that's valid for the new tenant.
-                        // The old refresh token is still associated with the old tenant, so
-                        // tenant-specific refresh will fail until we get new tokens.
-                        self.logger.info("Using standard OAuth refresh after server-side tenant switch to get new tokens for tenant: \(tenantId)")
-                        data = try await self.api.refreshToken(
-                            refreshToken: refreshToken,
-                            tenantId: nil
-                        )
-                        self.logger.info("Standard OAuth refresh successful after tenant switch. New tokens will be saved with tenant ID: \(tenantId)")
-                    } else {
-                        data = try await self.api.refreshToken(
-                            refreshToken: refreshToken,
-                            tenantId: nil
-                        )
-                    }
-                    
-                    self.logger.info("Token refresh successful, updating credentials")
-                    await self.setCredentials(accessToken: data.access_token, refreshToken: data.refresh_token)
-                    if let user = self.user {
-                        let newTenantId = user.activeTenant.id
-                        if newTenantId != tenantId {
-                            self.logger.warning("Tenant switch returned different tenant ID (\(newTenantId)) than expected (\(tenantId)). Updating stored tenant ID.")
-                            self.credentialManager.saveLastActiveTenantId(newTenantId)
-                        }
-                        self.logger.info("Tenant switch completed. New active tenant: \(user.activeTenant.name) (ID: \(newTenantId), tenantId: \(user.activeTenant.tenantId))")
-                        
-                        if newTenantId != tenantId && user.activeTenant.tenantId != tenantId {
-                            self.logger.warning("Tenant switch may have failed - expected \(tenantId) but got \(newTenantId)")
-                        }
-                        
-                        await MainActor.run {
-                            self.setIsLoading(false)
-                        }
-                        completion?(.success(user))
-                    } else {
-                        self.logger.error("User is nil after tenant switch and refresh")
-                        await MainActor.run {
-                            self.setIsLoading(false)
-                        }
-                        completion?(.failure(FronteggError.authError(.failedToSwitchTenant)))
-                    }
-                } catch {
-                    self.logger.error("Failed to refresh token after tenant switch: \(error)")
-                    await MainActor.run {
-                        self.setIsLoading(false)
-                    }
-                    completion?(.failure(FronteggError.authError(.failedToSwitchTenant)))
-                }
-                
-            }
-        }
-    }
-    
+    // MARK: Hosted Flows — see FronteggAuth+HostedFlows.swift
+    // MARK: Social Flows — see FronteggAuth+SocialFlows.swift
+    // MARK: Embedded & DeepLink — see FronteggAuth+EmbeddedAndDeepLink.swift
     // MARK: Advanced Auth — see FronteggAuth+AdvancedAuth.swift
 
 }
